app/requirements.txt
fastapi==0.96.0
numpy==1.23.1
loguru==0.7.0
torch==2.4.1
transformers==4.25.0
tqdm==4.66.5
matplotlib==3.5.2
scikit-learn==1.0.2
uvicorn==0.29.0
dotenv==1.0.1
pinecone==5.4.0
google-cloud-storage==2.18.2